+++ src/main/java/de/samply/frontend/dto/Project.java
@@ -1,6 +1,7 @@
 package de.samply.frontend.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import de.samply.annotations.IgnoreProjectConfigurationMatch;
 import de.samply.project.ProjectType;
 import de.samply.project.state.ProjectState;
 import de.samply.query.OutputFormat;
@@ -14,12 +15,18 @@ import java.time.LocalDate;
 @Data
 public class Project {
 
+    @IgnoreProjectConfigurationMatch
     private String code;
+    @IgnoreProjectConfigurationMatch
     private String creatorEmail;
+    @IgnoreProjectConfigurationMatch
     private String creatorName;
+    @IgnoreProjectConfigurationMatch
     private Instant createdAt;
     private LocalDate expiresAt;
+    @IgnoreProjectConfigurationMatch
     private Instant archivedAt;
+    @IgnoreProjectConfigurationMatch
     private Instant modifiedAt;
     private ProjectState state;
     private ProjectType type;
@@ -28,8 +35,11 @@ public class Project {
     private QueryFormat queryFormat;
     private OutputFormat outputFormat;
     private String templateId;
+    @IgnoreProjectConfigurationMatch
     private String label;
+    @IgnoreProjectConfigurationMatch
     private String description;
+    @IgnoreProjectConfigurationMatch
     private String explorerUrl;
     private String queryContext;
     private boolean isCustomConfig;